refactor(cli): move ls to a different file
